Move vault-plugin to components
