+++ components/vault-plugin/Dockerfile
