+++ components/vault-plugin/Gopkg.toml
