+++ components/vault-plugin/Makefile
@@ -1,3 +1,4 @@
+TSI_VERSION=$(shell cat ../../tsi-version.txt)
 TOOL?=vault-plugin-auth-ti-jwt
 TEST?=$$(go list ./... | grep -v /vendor/)
 EXTERNAL_TOOLS=\
@@ -12,7 +13,7 @@ BUILD_DATE="$(shell date -u +"%Y-%m-%dT%H:%M:%SZ")"
 IMG_NAME="ti-vault"
 REPO ?= trustedseriviceidentity
 IMAGE := $(REPO)/$(IMG_NAME):$(GIT_COMMIT_SHA)
-MUTABLE_IMAGE := $(REPO)/$(IMG_NAME):v0.4
+MUTABLE_IMAGE := $(REPO)/$(IMG_NAME):$(TSI_VERSION)
 
 ARCH=$(shell if test `uname -m` = "x86_64"; then echo amd64 ; else echo `uname -m`; fi;)
