+++ components/vault-plugin/cmd/vault-plugin-auth-ti-jwt/main.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	//"github.ibm.com/dpittner/vault-plugin-auth-ti-jwt"
-	"github.com/IBM/trusted-service-identity/examples/vault-plugin"
+	"github.com/IBM/trusted-service-identity/components/vault-plugin"
 	"github.com/hashicorp/vault/helper/pluginutil"
 	"github.com/hashicorp/vault/logical/plugin"
 )
