+++ components/vault-plugin/docker-vault/Dockerfile
