+++ components/vault-plugin/docker-vault/docker-entrypoint.sh
