+++ components/vault-plugin/local.json
