+++ components/vault-plugin/path_config.go
