+++ components/vault-plugin/path_config_test.go
