+++ components/vault-plugin/path_role.go
