+++ components/vault-plugin/scripts/build.sh
