+++ components/vault-plugin/scripts/update_deps.sh
